refactor: useFetch 커스텀훅 생성
1. TodoState에서 생성한 init함수를 인자로 넘겨준다
2. useFetch.js의 의미
- useFetch라는 함수를 가져와서 TodoState.js파일 안의 StateProvider 컴포넌트에서 useEffect 를 사용하는것. 즉 컴포넌트가 생성된다음에 useEffect가 일어나고 useFetch에서 useEffect가 포함되어있는이유는 불러진 컴포넌트에서 사용하라는뜻이다.
3. isLoading을 리턴한다.
